Added a HeaderIconButton for generic icons in the header. Removed HeaderAddButton and refactored its use to HeaderIconButton. Added a button to access passcode change UI. Upgraded
