+++ src/components/MainSettingsScreen/index.tsx
@@ -2,6 +2,5 @@
  * @module components
  */
 
-export { HeaderAddButton } from "./HeaderAddButton";
 export { ScheduleListHeader } from "./ScheduleListHeader";
 export { ScheduleListItem } from "./ScheduleListItem";
