+++ src/components/index.tsx
@@ -5,9 +5,10 @@
  * Complex items to be displayed
  */
 
-export { Clock, SlideUpIndicator, SnoozeButton } from "./HomeScreen";
-export { HeaderAddButton, ScheduleListHeader, ScheduleListItem } from "./MainSettingsScreen";
+export { Clock, Message, SlideUpIndicator, SnoozeButton } from "./HomeScreen";
+export { ScheduleListHeader, ScheduleListItem } from "./MainSettingsScreen";
 
+export { HeaderIconButton } from "./HeaderIconButton";
 export { HeaderBackButton } from "./HeaderBackButton";
 export { TabBarIcon } from "./TabBarIcon";
 export { ToggleButton } from "./ToggleButton";
